feat: ✨ 接入了AI五育分析接口&修改了表结构&修复了Excel导出数据相关bug
- 接入了五育分析接口，目前实现了个人诊断报告，班级诊断报告待实现；
- 个人诊断报告加入了导出为PDF功能，修改了样式，使其更加动态化；
- 修改了 wuyu_score 表结构，将分数字段由FLOAT类型改为INT类型；
- 新增了 wuyu_analysis_result表，用于存储诊断报告。

+++ ads-admin/src/main/java/io/ads/modules/analysis/excel/WuyuWeightExcel.java
@@ -6,7 +6,7 @@ import com.alibaba.excel.annotation.write.style.ContentRowHeight;
 import com.alibaba.excel.annotation.write.style.HeadRowHeight;
 import lombok.Data;
 
-import java.math.BigDecimal;
+
 import java.util.Date;
 
 /**
@@ -22,66 +22,66 @@ public class WuyuWeightExcel {
     @ExcelProperty(value = "学校ID")
     private Integer schoolId;
     @ExcelProperty(value = "德育")
-    private BigDecimal moral;
+    private Integer moral;
     @ExcelProperty(value = "智育")
-    private BigDecimal intellectual;
+    private Integer intellectual;
     @ExcelProperty(value = "体育")
-    private BigDecimal physical;
+    private Integer physical;
     @ExcelProperty(value = "美育")
-    private BigDecimal artistic;
+    private Integer artistic;
     @ExcelProperty(value = "劳育")
-    private BigDecimal labor;
+    private Integer labor;
     @ExcelProperty(value = "品德评定")
-    private BigDecimal characterEthics;
+    private Integer characterEthics;
     @ExcelProperty(value = "奖惩记录")
-    private BigDecimal rewardsPunishments;
+    private Integer rewardsPunishments;
     @ExcelProperty(value = "德育课程")
-    private BigDecimal moralEducationCourses;
+    private Integer moralEducationCourses;
     @ExcelProperty(value = "实践活动")
-    private BigDecimal practicalActivities;
+    private Integer practicalActivities;
     @ExcelProperty(value = "网络文化")
-    private BigDecimal onlineCulture;
+    private Integer onlineCulture;
     @ExcelProperty(value = "人际关系")
-    private BigDecimal interpersonalRelationships;
+    private Integer interpersonalRelationships;
     @ExcelProperty(value = "预习管理")
-    private BigDecimal prepManagement;
+    private Integer prepManagement;
     @ExcelProperty(value = "计划管理")
-    private BigDecimal planManagement;
+    private Integer planManagement;
     @ExcelProperty(value = "课堂行为")
-    private BigDecimal classroomBehavior;
+    private Integer classroomBehavior;
     @ExcelProperty(value = "课堂考勤")
-    private BigDecimal classroomAttendance;
+    private Integer classroomAttendance;
     @ExcelProperty(value = "作业管理")
-    private BigDecimal homeworkManagement;
+    private Integer homeworkManagement;
     @ExcelProperty(value = "复习管理")
-    private BigDecimal reviewManagement;
+    private Integer reviewManagement;
     @ExcelProperty(value = "个人能力")
-    private BigDecimal personalAbilities;
+    private Integer personalAbilities;
     @ExcelProperty(value = "学业成绩")
-    private BigDecimal academicPerformance;
+    private Integer academicPerformance;
     @ExcelProperty(value = "实验竞赛")
-    private BigDecimal experimentalCompetitions;
+    private Integer experimentalCompetitions;
     @ExcelProperty(value = "体检指标")
-    private BigDecimal examinationMetrics;
+    private Integer examinationMetrics;
     @ExcelProperty(value = "体能成绩")
-    private BigDecimal physicalFitnessScores;
+    private Integer physicalFitnessScores;
     @ExcelProperty(value = "体育特长")
-    private BigDecimal sportingSpecialties;
+    private Integer sportingSpecialties;
     @ExcelProperty(value = "健康生活")
-    private BigDecimal healthyLiving;
+    private Integer healthyLiving;
     @ExcelProperty(value = "心理素质")
-    private BigDecimal mentalQualities;
+    private Integer mentalQualities;
     @ExcelProperty(value = "体育课程")
-    private BigDecimal physicalEducationCourses;
+    private Integer physicalEducationCourses;
     @ExcelProperty(value = "美育课程")
-    private BigDecimal artsCourses;
+    private Integer artsCourses;
     @ExcelProperty(value = "美育成果")
-    private BigDecimal artsAchievements;
+    private Integer artsAchievements;
     @ExcelProperty(value = "美育活动")
-    private BigDecimal artsActivities;
+    private Integer artsActivities;
     @ExcelProperty(value = "劳动实践")
-    private BigDecimal laborPractices;
+    private Integer laborPractices;
     @ExcelProperty(value = "劳动课程")
-    private BigDecimal laborCourses;
+    private Integer laborCourses;
 
 }
